// src/app/shared/interfaces/restaurant.interface.ts
// Generated by https://quicktype.io

export interface Restaurant {
    restaurantName:        string;
    restaurantDescription: string;
    restaurantCategory:    string;
    restaurantCategoryArrayForm?: string[]
    restaurantCuisine:     string;
    restaurantCuisineArrayForm?: string[]
    restaurantImage:       string;
    isOpen:                boolean;
    id:                    number;
}


export interface allRestaurantsResponse {
    allRestaurants:Restaurant[];
}